feat: APIのサンプル作成
