+++ api/hello/go.mod
@@ -1,0 +1,10 @@
+module github.com/calmato/gran-book/api/hello
+
+go 1.15
+
+require (
+	github.com/golang/protobuf v1.4.3
+	google.golang.org/genproto v0.0.0-20201214200347-8c77b98c765d
+	google.golang.org/grpc v1.34.0
+	google.golang.org/protobuf v1.25.0
+)
