+++ infra/gateway/go.mod
@@ -3,9 +3,9 @@ module github.com/calmato/gran-book/infra/gateway
 go 1.15
 
 require (
+	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
 	github.com/golang/protobuf v1.4.3
 	github.com/grpc-ecosystem/grpc-gateway/v2 v2.0.1
-	github.com/iamrajiv/helloworld-grpc-gateway v0.0.0-20201214121557-ef10f17a0c91 // indirect
 	google.golang.org/genproto v0.0.0-20201204160425-06b3db808446
 	google.golang.org/grpc v1.34.0
 	google.golang.org/protobuf v1.25.0
